fix CClock and CWeather

+++ src/CWeather.cpp
@@ -1,8 +1,8 @@
 #include "CWeather.h"
 #include "memory.hpp"
 
-short &CWeather::Sandstorm = *(short *)(memory::GetAddr(0x50B3004));
-short &CWeather::UnderWaterness = *(short *)(memory::GetAddr(0x5571038));
+float &CWeather::Sandstorm = *(float *)(memory::GetAddr(0x50B3004));
+float &CWeather::UnderWaterness = *(float *)(memory::GetAddr(0x5571038));
 short &CWeather::ForcedWeatherType = *(short *)(memory::GetAddr(0x51425F0));
 short &CWeather::NewWeatherType = *(short *)(memory::GetAddr(0x50D5294));
 short &CWeather::OldWeatherType = *(short *)(memory::GetAddr(0x50D5298));
